backend/llama3-designer/requirements.txt
flask
flask-restful
flask-jwt-extended
pika  
python-dotenv  
llama-cpp-python
